requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
joblib==1.2.0
numpy==1.24.3
PyJWT==2.6.0
pytz==2022.7.1
scikit-learn==1.2.2
scipy==1.10.1
sqlparse==0.4.3
threadpoolctl==3.1.0
tzdata==2022.7
gunicorn